+++ src/index.ts
@@ -1,7 +1,7 @@
 import type { ESLint } from "eslint";
 import { readFile } from "node:fs/promises";
 import noDeprecatedCdsFioriDraftCompat from "./rules/no-deprecated-cds-fiori-draft-compat.js";
-import noDeprecatedCdsFeaturesOdataNewAdapter from "./rules/no-deprecated-cds-features-odata-new-adapter.js";
+import noDeprecatedCdsFeaturesOdataNewAdapter from "./rules/cds-features-odata-new-adapter.js";
 import cdsFeaturesCdsValidate from "./rules/cds-features-cds-validate.js";
 
 const pkg = JSON.parse(
@@ -14,7 +14,7 @@ const pkg = JSON.parse(
  */
 const rules = {
   "no-deprecated-cds-fiori-draft-compat": noDeprecatedCdsFioriDraftCompat,
-  "no-deprecated-cds-features-odata-new-adapter":
+  "cds-features-odata-new-adapter":
     noDeprecatedCdsFeaturesOdataNewAdapter,
   "cds-features-cds-validate": cdsFeaturesCdsValidate,
 };
@@ -35,7 +35,7 @@ const plugin: ESLint.Plugin = {
       },
       rules: {
         "cds-deprecated-config/no-deprecated-cds-fiori-draft-compat": "error",
-        "cds-deprecated-config/no-deprecated-cds-features-odata-new-adapter":
+        "cds-deprecated-config/cds-features-odata-new-adapter":
           "error",
         "cds-deprecated-config/cds-features-cds-validate":
           "error",